backend/requirements.txt
uvicorn==0.30.6
fastapi==0.115.0
selenium
requests==2.32.5
sqlalchemy==2.0.35
httpx==0.28.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client
gspread==6.2.1
python-dotenv==1.0.1
cryptography==43.0.3
pandas==2.3.3
openpyxl==3.1.5
python-multipart==0.0.9
pyjwt==2.11.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
